requirements.txt
aiohttp==3.8.3
aiohttp-socks==0.8.0
python-socks==2.0.3
uvicorn==0.18.3
fastapi==0.88.0
nats-py==2.2.0

redis==4.3.4
aiochclient==2.2.0
clickhouse-driver==0.2.5

pandas==1.5.2
numpy==1.24.0
pyarrow==10.0.1

prometheus-client==0.15.0

pyyaml==6.0
python-dotenv==0.21.0

structlog==22.1.0
colorama==0.4.6
psutil==5.9.4

# 定时任务依赖
croniter==1.3.15 